+++ simvue/api/objects/administrator/tenant.py
@@ -15,11 +15,13 @@ try:
 except ImportError:
     from typing_extensions import override  # noqa: UP035
 
+import datetime
 import typing
 
 import pydantic
 
 from simvue.api.objects.base import SimvueObject, staging_check, write_only
+from simvue.models import DATETIME_FORMAT
 
 
 class Tenant(SimvueObject):
@@ -137,3 +139,16 @@ class Tenant(SimvueObject):
     def max_data_volume(self, max_data_volume: int) -> None:
         """Update tenant's maximum data volume."""
         self._staging["max_data_volume"] = max_data_volume
+
+    @property
+    def created(self) -> datetime.datetime | None:
+        """Set/retrieve created datetime for the run.
+
+        Returns
+        -------
+        datetime.datetime
+        """
+        _created: str | None = self._get_attribute("created")
+        return (
+            datetime.datetime.strptime(_created, DATETIME_FORMAT) if _created else None
+        )